Move clog .NET generation to its own high level cmake target
This makes it to the global CLOG dependency only runs once. It also simplifies the targets for the individual sources.

Also updates the name, as CLOG_ADD_SOURCEFILE wasn't technically correct.

Will require adding the inc subdirectory in the calling project, but it makes the build much faster, much easier, and cleaner.